[fix/#5]: 위도, 경도 타입 변경

+++ src/test/java/fisa/woorizip/backend/support/fixture/FacilityFixture.java
@@ -12,8 +12,8 @@ public class FacilityFixture {
     private String name = "이가네 짜글이";
     private String gu = "마포구";
     private String dong = "상암동";
-    private BigDecimal latitude = BigDecimal.valueOf(37.5813506);
-    private BigDecimal longitude = BigDecimal.valueOf(126.886193);
+    private double latitude = 37.5813506;
+    private double longitude = 126.886193;
 
     public static FacilityFixture builder() {
         return new FacilityFixture();
@@ -44,12 +44,12 @@ public class FacilityFixture {
         return this;
     }
 
-    public FacilityFixture latitude(BigDecimal latitude) {
+    public FacilityFixture latitude(double latitude) {
         this.latitude = latitude;
         return this;
     }
 
-    public FacilityFixture longitude(BigDecimal longitude) {
+    public FacilityFixture longitude(double longitude) {
         this.longitude = longitude;
         return this;
     }
